requirements.txt
flask-cors
flask_seeder
passlib
python-decouple
flask_script
flask_migrate
python-dotenv
flask-bcrypt
authlib==0.14.3
flask-mongoengine
blinker==1.4
flask-fs==0.6.1
Flask-Caching==1.9.0
Flask-Navigation==0.2.0
speaklater==1.3
werkzeug==0.16.1 
Pillow==8.0.1
slugify==0.0.1
factory-boy==3.2.0
netaddr==0.8.0
kombu==5.0.2
tlds==2020122800
Flask-Security==3.0.0
elasticsearch-dsl==7.3.0
redis==2.6.2
flask-caching==1.9.0
flask_wtf
